WEBC-340: Moved BadRequest exception type and the Request VO into "Request" namespace and moved some general request parameter parsing method into Request\Request VO class
