+++ src/Service/Authenticator/PublicAccess.php
@@ -25,7 +25,7 @@ use Shopgate\CloudIntegrationSdk\ValueObject\Request;
 
 class PublicAccess implements AuthenticatorInterface
 {
-    public function authenticate(Request $request)
+    public function authenticate(Request\Request $request)
     {
         // no authentication required
     }
