+++ src/Service/RequestHandler/PostAuthToken.php
@@ -54,7 +54,7 @@ class PostAuthToken implements RequestHandlerInterface
         return $this->authenticator;
     }
 
-    public function handle(Request $request)
+    public function handle(Request\Request $request)
     {
         return new Response(200, [], ""); // TODO: Implement handle() method.
     }
